fix  adding interface page
make the LGTM bot happy by using 'new'
also use UI instead of AB so things match better

check if isValid is a function before trying to use it -- thi sspecifically broke adding a page to the root level OR to a child of root. Adding a page on the  ./../../___ level worked sometimes... for some reason?